navigation draweré

+++ lib/service/auth_service.dart
@@ -1,6 +1,10 @@
 import 'dart:convert';
 import 'package:http/http.dart' as http;
+import 'package:http_interceptor/http/intercepted_client.dart';
 import 'package:madaride/utils/file_secure_storage.dart';
+
+import '../model/user.dart';
+import '../utils/http_interceptor.dart';
 
 class AuthService {
   final fileStorage = SecureStorage();
@@ -30,17 +34,57 @@ class AuthService {
     return response.statusCode == 201;
   }
 
-  Future<void> logout() async {
-    await fileStorage.deleteToken();
+
+
+  Future<User> profile() async {
+    final client = InterceptedClient.build(
+      interceptors: [AuthInterceptor()],
+    );
+
+    final response = await client.get(
+      Uri.parse('$baseUrl/profile'),
+      headers: <String, String>{
+        'Content-Type': 'application/json; charset=UTF-8',
+      },
+    );
+
+    if (response.statusCode == 200) {
+      Map<String, dynamic> body = jsonDecode(response.body);
+      return User.fromJson(body);
+    } else {
+      throw Exception('Failed to load user: ${response.statusCode}');
+    }
+  }
+
+  Future<bool> logout() async {
+    final client = InterceptedClient.build(
+      interceptors: [AuthInterceptor()],
+    );
+
+    final response = await client.post(
+      Uri.parse('$baseUrl/logout'),
+      body: {},
+    );
+
+    if (response.statusCode == 200) {
+      await fileStorage.deleteToken();
+      return true;
+    }
+
+    return false;
   }
 
   Future<String?> refreshToken() async {
     String? oldToken = fileStorage.getToken() as String?;
     if (oldToken == null) return null;
 
+    final client = InterceptedClient.build(
+      interceptors: [AuthInterceptor()],
+    );
+
     try {
-      final response = await http.post(
-        Uri.parse('$baseUrl/refresh-token'),
+      final response = await client.post(
+        Uri.parse('$baseUrl/refresh'),
         headers: {'Authorization': 'Bearer $oldToken'},
       );
